Safari-611.1.10/LayoutTests/fast/canvas/reset-scaling-by-height-change-expected.txt

The test to ensure resetting scaling by height change

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS false is false
PASS false is false
PASS successfullyParsed is true

TEST COMPLETE
